% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{istd}
\alias{istd}
\title{Example for data_expenses object}
\format{
10 rows, 7 columns. Columns are object, real_projected (Real/Projected), paid_pending (Paid/Pending), description, society, paymaster, amount
}
\usage{
data("istd")
}
\description{
This data set is a small example.
}
\examples{
data("istd")

}
\author{
Margot Brard \email{margot@thinkr.fr}
}
\keyword{datasets}
